#58 change file CardManager class
